refactor(codex): split stateful projector owners
