+++ extensions/codex/src/app-server/event-projector-usage.ts
@@ -1,0 +1,20 @@
+import { normalizeUsage } from "openclaw/plugin-sdk/agent-harness-runtime";
+import { readNumber } from "./event-projector-values.js";
+import type { JsonObject } from "./protocol.js";
+
+export function normalizeCodexTokenUsage(record: JsonObject): ReturnType<typeof normalizeUsage> {
+  // v2 TokenUsageBreakdown. inputTokens includes cached input; OpenClaw usage
+  // tracks uncached input and cache reads separately.
+  const inputTokens = readNumber(record, "inputTokens");
+  const cacheRead = readNumber(record, "cachedInputTokens");
+  const input =
+    inputTokens !== undefined && cacheRead !== undefined
+      ? Math.max(0, inputTokens - cacheRead)
+      : inputTokens;
+  return normalizeUsage({
+    input,
+    output: readNumber(record, "outputTokens"),
+    cacheRead,
+    total: readNumber(record, "totalTokens"),
+  });
+}
