+++ extensions/codex/src/app-server/event-projector.ts
@@ -3,75 +3,35 @@ import {
   classifyAgentHarnessTerminalOutcome,
   embeddedAgentLog,
   emitAgentEvent as emitGlobalAgentEvent,
-  inferToolMetaFromArgs,
-  normalizeUsage,
   runAgentHarnessAfterCompactionHook,
-  runAgentHarnessAfterToolCallHook,
   runAgentHarnessBeforeCompactionHook,
-  TOOL_PROGRESS_OUTPUT_MAX_CHARS,
-  type AgentMessage,
   type BeforeToolCallFailureDisposition,
   type EmbeddedRunAttemptParams,
   type EmbeddedRunAttemptResult,
   type HeartbeatToolResponse,
   type MessagingToolSend,
   type MessagingToolSourceReplyPayload,
-  type ToolProgressDetailMode,
 } from "openclaw/plugin-sdk/agent-harness-runtime";
-import { generatedImageAssetFromBase64 } from "openclaw/plugin-sdk/image-generation";
-import type { Usage } from "openclaw/plugin-sdk/llm";
-import { saveMediaBuffer } from "openclaw/plugin-sdk/media-store";
-import { asDateTimestampMs } from "openclaw/plugin-sdk/number-runtime";
-import { truncateUtf16Safe } from "openclaw/plugin-sdk/text-utility-runtime";
 import { CodexAssistantProjection } from "./event-projector-assistant.js";
+import { CodexEventProjection } from "./event-projector-events.js";
 import {
-  isMutatingNativeToolItem,
-  isNonSuccessItemStatus,
-  isSideEffectingNativeToolItem,
-  itemKind,
   itemName,
   itemStatus,
-  itemTitle,
   shouldClearTerminalPresentationForNativeItem,
-  shouldRecordNativeToolTranscript,
   shouldSynthesizeToolProgressForItem,
 } from "./event-projector-items.js";
+import { CodexGeneratedMediaProjection } from "./event-projector-media.js";
 import { CodexNativeToolLifecycleProjector } from "./event-projector-native-tool-lifecycle.js";
 import { CodexReasoningProjection } from "./event-projector-reasoning.js";
-import {
-  isNativePostToolUseRelayItem,
-  itemMeta,
-  itemOutputText,
-  itemToolArgs,
-  itemToolError,
-  itemToolResult,
-  itemTranscriptResultText,
-  nativeToolActionFingerprint,
-  shouldSuppressChannelProgressForItem,
-} from "./event-projector-tool-items.js";
-import {
-  collectDynamicToolContentText,
-  formatToolOutput,
-  formatToolSummary,
-  MAX_TOOL_OUTPUT_DELTA_MESSAGES_PER_ITEM,
-  normalizeToolTranscriptArguments,
-  TOOL_PROGRESS_ECHO_PREFIX_MIN_CHARS,
-  TOOL_PROGRESS_ECHO_SIGNATURE_CAP,
-  TOOL_TRANSCRIPT_OUTPUT_MAX_CHARS,
-  ToolOutputAccumulator,
-  toolOutputRawEchoSignature,
-  truncateToolTranscriptText,
-} from "./event-projector-tool-output.js";
+import { CodexToolProgressProjection } from "./event-projector-tool-progress.js";
+import { CodexToolTranscriptProjection } from "./event-projector-tool-transcript.js";
+import { normalizeCodexTokenUsage } from "./event-projector-usage.js";
 import {
   readCodexErrorNotificationMessage,
-  readHookOutputEntries,
   readItem,
   readItemString,
-  readNullableString,
-  readNumber,
   readString,
 } from "./event-projector-values.js";
-import { resolveCodexLocalRuntimeAttribution } from "./local-runtime-attribution.js";
 import type { CodexNativePreToolUseFailure } from "./native-hook-relay.js";
 import {
   readCodexNotificationThreadId,
@@ -88,16 +48,12 @@ import {
   type JsonValue,
 } from "./protocol.js";
 import { formatCodexUsageLimitErrorMessage } from "./rate-limits.js";
-import { readCodexMirroredSessionHistoryMessages } from "./session-history.js";
-import {
-  resolveCodexToolProgressDetailMode,
-  sanitizeCodexToolArguments,
-} from "./tool-progress-normalization.js";
 import type { CodexTrajectoryRecorder } from "./trajectory.js";
 import { attachCodexMirrorIdentity } from "./upstream-prompt-provenance.js";
 import { promptSnapshot } from "./user-prompt-message.js";
 
 export { CodexNativeToolLifecycleProjector };
+export { shouldEmitTranscriptToolProgress } from "./event-projector-tool-progress.js";
 
 type CodexAppServerToolTelemetry = {
   didSendViaMessagingTool: boolean;
@@ -122,116 +78,25 @@ type CodexAppServerEventProjectorOptions = {
   upstreamUserText?: string;
 };
 
-const ZERO_USAGE: Usage = {
-  input: 0,
-  output: 0,
-  cacheRead: 0,
-  cacheWrite: 0,
-  totalTokens: 0,
-  cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 },
-};
-
-// FIFO holds genuinely distinct emitted shapes (summary/chunk/final/aggregate). Stream
-// accumulation owns a dedicated slot and does not consume FIFO capacity.
-const MISSING_TOOL_RESULT_ERROR =
-  "OpenClaw recorded a native Codex tool.call without a matching tool.result before the turn completed.";
-const GENERATED_IMAGE_MEDIA_SUBDIR = "tool-image-generation";
-
-function formatMissingToolResultError(params: { id: string; name: string }): string {
-  return `${MISSING_TOOL_RESULT_ERROR} toolCallId=${params.id}; toolName=${params.name}`;
-}
-const BYTES_PER_MB = 1024 * 1024;
-// Match OpenClaw's default image media cap for generated image tool outputs.
-const DEFAULT_GENERATED_IMAGE_MAX_BYTES = 6 * BYTES_PER_MB;
-const TRANSCRIPT_PROGRESS_SUPPRESSED_TOOL_NAMES = new Set([
-  "message",
-  "messages",
-  "reply",
-  "send",
-  "reaction",
-  "react",
-  "typing",
-]);
-
-export function shouldEmitTranscriptToolProgress(toolName: unknown, _args?: unknown): boolean {
-  const normalized = typeof toolName === "string" ? toolName.trim().toLowerCase() : "";
-  return Boolean(normalized && !TRANSCRIPT_PROGRESS_SUPPRESSED_TOOL_NAMES.has(normalized));
-}
-
-type ToolTranscriptCallInput = {
-  id: string;
-  name: string;
-  arguments?: unknown;
-};
-
-type ToolTranscriptResultInput = {
-  id: string;
-  name: string;
-  text?: string;
-  isError: boolean;
-};
-
-function toolResultStatusText(params: ToolTranscriptResultInput): string {
-  return params.isError ? `${params.name} failed` : `${params.name} completed`;
-}
-
-type ToolProgressRawSignature = {
-  length: number;
-  prefix: string;
-};
-
-type ToolProgressEchoState = {
-  displayTexts: string[];
-  // Single slot for handleOutputDelta accumulation; replaced per delta (not appended).
-  streamedDisplayText?: string;
-  // One logical stream shape; replaced per delta (not pushed into rawSignatures FIFO).
-  streamedRawSignature?: ToolProgressRawSignature;
-  rawSignatures: ToolProgressRawSignature[];
-};
-
 export class CodexAppServerEventProjector {
   private readonly assistantProjection: CodexAssistantProjection;
   private readonly reasoningProjection: CodexReasoningProjection;
   private readonly activeItemIds = new Set<string>();
   private readonly completedItemIds = new Set<string>();
   private readonly activeCompactionItemIds = new Set<string>();
-  private readonly toolProgressEchoesByItem = new Map<string, ToolProgressEchoState>();
-  private readonly toolResultSummaryItemIds = new Set<string>();
-  private readonly toolResultOutputItemIds = new Set<string>();
-  private readonly toolResultOutputStreamedItemIds = new Set<string>();
-  private readonly transcriptToolProgressSuppressedIds = new Set<string>();
-  private readonly toolTranscriptArgumentsById = new Map<string, unknown>();
-  private readonly toolResultOutputDeltaState = new Map<
-    string,
-    { chars: number; messages: number; truncated: boolean }
-  >();
-  private readonly toolOutput = new ToolOutputAccumulator();
-  private readonly toolMetas = new Map<string, EmbeddedRunAttemptResult["toolMetas"][number]>();
   private readonly terminalPresentationClearedItemIds = new Set<string>();
   private readonly nativeToolOutcomeOrdinals = new Map<string, number>();
-  private readonly sideEffectingToolItemIds = new Set<string>();
-  private readonly sideEffectingDynamicToolCallIds = new Set<string>();
-  private readonly toolTranscriptMessages: AgentMessage[] = [];
-  private readonly toolTranscriptCallIds = new Set<string>();
-  private readonly toolTranscriptResultIds = new Set<string>();
-  private readonly toolTranscriptNamesById = new Map<string, string>();
-  private readonly toolTrajectoryCallIds = new Set<string>();
-  private readonly toolTrajectoryResultIds = new Set<string>();
-  private readonly toolTrajectoryNamesById = new Map<string, string>();
-  private readonly toolTrajectoryItemsById = new Map<string, CodexThreadItem>();
-  private readonly transcriptToolProgressCallIds = new Set<string>();
-  private lastNativeToolError: EmbeddedRunAttemptResult["lastToolError"];
-  private readonly nativeGeneratedMediaItemIds = new Set<string>();
-  private readonly nativeGeneratedMediaUrlsByItemId = new Map<string, string>();
+  private readonly generatedMediaProjection: CodexGeneratedMediaProjection;
+  private readonly eventProjection: CodexEventProjection;
   private readonly nativeToolLifecycleProjector: CodexNativeToolLifecycleProjector;
-  private readonly afterToolCallObservedItemIds = new Set<string>();
+  private readonly toolProgressProjection: CodexToolProgressProjection;
+  private readonly toolTranscriptProjection: CodexToolTranscriptProjection;
   private completedTurn: CodexTurn | undefined;
   private promptError: unknown;
   private promptErrorSource: EmbeddedRunAttemptResult["promptErrorSource"] = null;
   private synthesizedMissingToolResultError: string | null = null;
   private aborted = false;
-  private tokenUsage: ReturnType<typeof normalizeUsage>;
-  private guardianReviewCount = 0;
+  private tokenUsage: ReturnType<typeof normalizeCodexTokenUsage>;
   private completedCompactionCount = 0;
 
   constructor(
@@ -248,10 +113,30 @@ export class CodexAppServerEventProjector {
         runAbortSignal: options.runAbortSignal,
       },
     );
+    this.generatedMediaProjection = new CodexGeneratedMediaProjection(params.config);
+    this.toolProgressProjection = new CodexToolProgressProjection(params);
+    this.toolTranscriptProjection = new CodexToolTranscriptProjection(
+      params,
+      threadId,
+      turnId,
+      this.toolProgressProjection,
+      {
+        nativePostToolUseRelayEnabled: options.nativePostToolUseRelayEnabled,
+        trajectoryRecorder: options.trajectoryRecorder,
+      },
+    );
+    this.eventProjection = new CodexEventProjection(
+      threadId,
+      turnId,
+      (event) => this.emitAgentEvent(event),
+      this.toolProgressProjection,
+      this.toolTranscriptProjection,
+      options.onNativeToolResultRecorded,
+    );
     this.assistantProjection = new CodexAssistantProjection(
       params,
       (event) => this.emitAgentEvent(event),
-      (text) => this.matchesToolProgressEcho(text),
+      (text) => this.toolProgressProjection.matchesEcho(text),
     );
     this.reasoningProjection = new CodexReasoningProjection(params, (event) =>
       this.emitAgentEvent(event),
@@ -359,18 +244,18 @@ export class CodexAppServerEventProjector {
         await this.handleItemCompleted(params);
         break;
       case "item/commandExecution/outputDelta":
-        this.handleOutputDelta(params, "bash");
+        this.toolProgressProjection.handleOutputDelta(params, "bash");
         break;
       case "item/autoApprovalReview/started":
       case "item/autoApprovalReview/completed":
-        this.handleGuardianReviewNotification(notification.method, params);
+        this.eventProjection.handleGuardianReview(notification.method, params);
         break;
       case "guardianWarning":
-        this.handleGuardianWarning(params);
+        this.eventProjection.handleGuardianWarning(params);
         break;
       case "hook/started":
       case "hook/completed":
-        this.handleHookNotification(notification.method, params);
+        this.eventProjection.handleHook(notification.method, params);
         break;
       case "thread/tokenUsage/updated":
         this.handleTokenUsage(params);
@@ -409,11 +294,16 @@ export class CodexAppServerEventProjector {
     const hasDeliverableAssistantOnCompletedTurn =
       this.completedTurn?.status === "completed" &&
       assistantTexts.some((text) => text.trim().length > 0);
-    this.synthesizeMissingToolResults({
-      synthesize: legacyFailClosed,
-      recordPromptError:
-        legacyFailClosed && !hasDeliverableAssistantOnCompletedTurn && !this.aborted,
-    });
+    const synthesizedMissingToolResultError =
+      this.toolTranscriptProjection.synthesizeMissingToolResults({
+        synthesize: legacyFailClosed,
+        recordPromptError:
+          legacyFailClosed && !hasDeliverableAssistantOnCompletedTurn && !this.aborted,
+      });
+    if (synthesizedMissingToolResultError) {
+      this.synthesizedMissingToolResultError = synthesizedMissingToolResultError;
+      this.promptErrorSource = this.promptErrorSource ?? "prompt";
+    }
     const assistantMessageOptions = {
       tokenUsage: this.tokenUsage,
       aborted: this.aborted,
@@ -457,7 +347,7 @@ export class CodexAppServerEventProjector {
         ),
       );
     }
-    messagesSnapshot.push(...this.toolTranscriptMessages);
+    messagesSnapshot.push(...this.toolTranscriptProjection.transcriptMessages);
     if (lastAssistant) {
       messagesSnapshot.push(attachCodexMirrorIdentity(lastAssistant, `${turnId}:assistant`));
     }
@@ -473,13 +363,12 @@ export class CodexAppServerEventProjector {
       promptError,
       turnCompleted: Boolean(this.completedTurn),
     });
-    const toolMetas = [...this.toolMetas.values()];
+    const toolMetas = this.toolProgressProjection.toolMetas;
     const hadPotentialSideEffects =
       toolTelemetry.didSendViaMessagingTool ||
       (toolTelemetry.successfulCronAdds ?? 0) > 0 ||
-      this.nativeGeneratedMediaItemIds.size > 0 ||
-      this.sideEffectingToolItemIds.size > 0 ||
-      this.sideEffectingDynamicToolCallIds.size > 0;
+      this.generatedMediaProjection.hasGeneratedMedia() ||
+      this.toolProgressProjection.hasPotentialSideEffects;
     return {
       aborted: this.aborted,
       externalAbort: false,
@@ -498,7 +387,9 @@ export class CodexAppServerEventProjector {
       toolMetas,
       lastAssistant,
       currentAttemptAssistant,
-      ...(this.lastNativeToolError ? { lastToolError: this.lastNativeToolError } : {}),
+      ...(this.toolProgressProjection.lastToolError
+        ? { lastToolError: this.toolProgressProjection.lastToolError }
+        : {}),
       didSendViaMessagingTool: toolTelemetry.didSendViaMessagingTool,
       didDeliverSourceReplyViaMessageTool:
         toolTelemetry.didDeliverSourceReplyViaMessageTool === true,
@@ -507,7 +398,7 @@ export class CodexAppServerEventProjector {
       messagingToolSentTargets: toolTelemetry.messagingToolSentTargets,
       messagingToolSourceReplyPayloads: toolTelemetry.messagingToolSourceReplyPayloads ?? [],
       heartbeatToolResponse: toolTelemetry.heartbeatToolResponse,
-      toolMediaUrls: this.buildToolMediaUrls(toolTelemetry),
+      toolMediaUrls: this.generatedMediaProjection.buildToolMediaUrls(toolTelemetry),
       toolAudioAsVoice: toolTelemetry.toolAudioAsVoice,
       successfulCronAdds: toolTelemetry.successfulCronAdds,
       cloudCodeAssistFormatError: false,
@@ -525,17 +416,13 @@ export class CodexAppServerEventProjector {
           : {}),
       },
       yieldDetected: options?.yieldDetected || false,
-      didSendDeterministicApprovalPrompt: this.guardianReviewCount > 0 ? false : undefined,
+      didSendDeterministicApprovalPrompt:
+        this.eventProjection.guardianReviewCount > 0 ? false : undefined,
     };
   }
 
   recordDynamicToolCall(params: { callId: string; tool: string; arguments?: JsonValue }): void {
-    const args = sanitizeCodexToolArguments(params.arguments);
-    this.recordToolTranscriptCall({
-      id: params.callId,
-      name: params.tool,
-      arguments: args,
-    });
+    this.toolTranscriptProjection.recordDynamicToolCall(params);
   }
 
   recordDynamicToolResult(params: {
@@ -547,35 +434,8 @@ export class CodexAppServerEventProjector {
     sideEffectEvidence?: boolean;
     contentItems: CodexDynamicToolCallOutputContentItem[];
   }): void {
-    const resultText = collectDynamicToolContentText(params.contentItems);
-    const existing = this.toolMetas.get(params.callId);
-    this.toolMetas.set(params.callId, {
-      toolName: existing?.toolName ?? params.tool,
-      ...(existing?.meta ? { meta: existing.meta } : {}),
-      ...(params.asyncStarted === true ? { asyncStarted: true } : {}),
-      ...(!params.success ? { isError: true } : {}),
-    });
-    this.recordToolTranscriptResult({
-      id: params.callId,
-      name: params.tool,
-      text: resultText,
-      isError: !params.success,
-    });
-    if (!params.success && params.terminalType === "blocked") {
-      this.lastNativeToolError = {
-        toolName: params.tool,
-        error: resultText || "codex dynamic tool blocked",
-      };
-    } else if (
-      params.success &&
-      this.lastNativeToolError &&
-      !this.lastNativeToolError.mutatingAction
-    ) {
-      this.lastNativeToolError = undefined;
-    }
-    if (params.sideEffectEvidence === true) {
-      this.sideEffectingDynamicToolCallIds.add(params.callId);
-    }
+    this.toolProgressProjection.recordDynamicToolResult(params);
+    this.toolTranscriptProjection.recordDynamicToolResult(params);
   }
 
   markTimedOut(): void {
@@ -604,7 +464,7 @@ export class CodexAppServerEventProjector {
       this.activeCompactionItemIds.add(itemId);
       await runAgentHarnessBeforeCompactionHook({
         sessionFile: this.params.sessionFile,
-        messages: await this.readMirroredSessionMessages(),
+        messages: await this.toolTranscriptProjection.readMirroredSessionMessages(),
         ctx: {
           runId: this.params.runId,
           agentId: this.params.agentId,
@@ -627,11 +487,11 @@ export class CodexAppServerEventProjector {
         },
       });
     }
-    this.recordToolMeta(item);
-    this.emitStandardItemEvent({ phase: "start", item });
-    await this.emitNormalizedToolItemEvent({ phase: "start", item });
-    this.recordNativeToolTranscriptCall(item);
-    this.emitToolResultSummary(item);
+    this.toolProgressProjection.recordToolMeta(item);
+    this.eventProjection.emitStandardItemEvent({ phase: "start", item });
+    await this.eventProjection.emitNormalizedToolItemEvent({ phase: "start", item });
+    this.toolTranscriptProjection.recordNativeToolCall(item);
+    this.toolProgressProjection.emitToolResultSummary(item);
     this.emitAgentEvent({
       stream: "codex_app_server.item",
       data: { phase: "started", itemId, type: item?.type },
@@ -649,14 +509,14 @@ export class CodexAppServerEventProjector {
     }
     this.assistantProjection.recordItemCompleted(item, itemId, this.activeItemIds);
     this.reasoningProjection.recordItem(item);
-    this.recordNativeGeneratedMedia(item);
+    this.generatedMediaProjection.recordNative(item);
     if (item?.type === "contextCompaction" && itemId) {
       this.activeCompactionItemIds.delete(itemId);
       this.completedCompactionCount += 1;
       this.options.onContextCompacted?.();
       await runAgentHarnessAfterCompactionHook({
         sessionFile: this.params.sessionFile,
-        messages: await this.readMirroredSessionMessages(),
+        messages: await this.toolTranscriptProjection.readMirroredSessionMessages(),
         compactedCount: -1,
         ctx: {
           runId: this.params.runId,
@@ -681,14 +541,14 @@ export class CodexAppServerEventProjector {
         },
       });
     }
-    this.recordToolMeta(item);
-    this.rememberCommandAggregateOutputEcho(item);
-    this.emitStandardItemEvent({ phase: "end", item });
-    await this.emitNormalizedToolItemEvent({ phase: "result", item });
-    this.recordNativeToolTranscriptCall(item);
-    this.recordNativeToolTranscriptResult(item);
-    this.emitToolResultSummary(item);
-    this.emitToolResultOutput(item);
+    this.toolProgressProjection.recordToolMeta(item);
+    this.toolProgressProjection.rememberCommandAggregateOutputEcho(item);
+    this.eventProjection.emitStandardItemEvent({ phase: "end", item });
+    await this.eventProjection.emitNormalizedToolItemEvent({ phase: "result", item });
+    this.toolTranscriptProjection.recordNativeToolCall(item);
+    this.toolTranscriptProjection.recordNativeToolResult(item);
+    this.toolProgressProjection.emitToolResultSummary(item);
+    this.toolProgressProjection.emitToolResultOutput(item);
     this.emitAgentEvent({
       stream: "codex_app_server.item",
       data: { phase: "completed", itemId, type: item?.type },
@@ -706,66 +566,6 @@ export class CodexAppServerEventProjector {
     if (usage) {
       this.tokenUsage = usage;
     }
-  }
-
-  private handleGuardianReviewNotification(method: string, params: JsonObject): void {
-    this.guardianReviewCount += 1;
-    const review = isJsonObject(params.review) ? params.review : undefined;
-    const action = isJsonObject(params.action) ? params.action : undefined;
-    this.emitAgentEvent({
-      stream: "codex_app_server.guardian",
-      data: {
-        method,
-        phase: method.endsWith("/started") ? "started" : "completed",
-        reviewId: readString(params, "reviewId"),
-        targetItemId: readNullableString(params, "targetItemId"),
-        decisionSource: readString(params, "decisionSource"),
-        status: review ? readString(review, "status") : undefined,
-        riskLevel: review ? readString(review, "riskLevel") : undefined,
-        userAuthorization: review ? readString(review, "userAuthorization") : undefined,
-        rationale: review ? readNullableString(review, "rationale") : undefined,
-        actionType: action ? readString(action, "type") : undefined,
-      },
-    });
-  }
-
-  private handleGuardianWarning(params: JsonObject): void {
-    this.emitAgentEvent({
-      stream: "codex_app_server.guardian",
-      data: {
-        phase: "warning",
-        message: readString(params, "message"),
-      },
-    });
-  }
-
-  private handleHookNotification(method: string, params: JsonObject): void {
-    const run = isJsonObject(params.run) ? params.run : undefined;
-    if (!run) {
-      return;
-    }
-    const durationMs = readNumber(run, "durationMs");
-    const entries = readHookOutputEntries(run.entries);
-    const hookTurnId = readNullableString(params, "turnId");
-    this.emitAgentEvent({
-      stream: "codex_app_server.hook",
-      data: {
-        phase: method === "hook/started" ? "started" : "completed",
-        threadId: this.threadId,
-        turnId: hookTurnId === undefined ? this.turnId : hookTurnId,
-        hookRunId: readString(run, "id"),
-        eventName: readString(run, "eventName"),
-        handlerType: readString(run, "handlerType"),
-        executionMode: readString(run, "executionMode"),
-        scope: readString(run, "scope"),
-        source: readString(run, "source"),
-        sourcePath: readString(run, "sourcePath"),
-        status: readString(run, "status"),
-        statusMessage: readNullableString(run, "statusMessage"),
-        ...(durationMs !== undefined ? { durationMs } : {}),
-        ...(entries.length > 0 ? { entries } : {}),
-      },
-    });
   }
 
   private async handleTurnCompleted(params: JsonObject): Promise<void> {
@@ -804,15 +604,15 @@ export class CodexAppServerEventProjector {
     for (const item of turnItems) {
       this.assistantProjection.recordSnapshotItem(item);
       this.reasoningProjection.recordItem(item);
-      this.recordNativeGeneratedMedia(item);
-      this.recordToolMeta(item);
-      this.rememberCommandAggregateOutputEcho(item);
+      this.generatedMediaProjection.recordNative(item);
+      this.toolProgressProjection.recordToolMeta(item);
+      this.toolProgressProjection.rememberCommandAggregateOutputEcho(item);
       await this.emitSnapshotOnlyNativeToolProgress(item);
-      this.recordNativeToolTranscriptCall(item);
-      this.recordNativeToolTranscriptResult(item);
-      this.emitAfterToolCallObservation(item);
-      this.emitToolResultSummary(item);
-      this.emitToolResultOutput(item);
+      this.toolTranscriptProjection.recordNativeToolCall(item);
+      this.toolTranscriptProjection.recordNativeToolResult(item);
+      this.toolTranscriptProjection.emitAfterToolCallObservation(item);
+      this.toolProgressProjection.emitToolResultSummary(item);
+      this.toolProgressProjection.emitToolResultOutput(item);
     }
     this.activeCompactionItemIds.clear();
     await this.reasoningProjection.maybeEndReasoning();
@@ -829,81 +629,18 @@ export class CodexAppServerEventProjector {
     }
     const wasStarted = this.activeItemIds.has(item.id);
     if (!wasStarted) {
-      this.emitStandardItemEvent({ phase: "start", item });
-      await this.emitNormalizedToolItemEvent({ phase: "start", item });
+      this.eventProjection.emitStandardItemEvent({ phase: "start", item });
+      await this.eventProjection.emitNormalizedToolItemEvent({ phase: "start", item });
     }
     this.activeItemIds.delete(item.id);
-    this.emitStandardItemEvent({ phase: "end", item });
-    await this.emitNormalizedToolItemEvent({ phase: "result", item });
+    this.eventProjection.emitStandardItemEvent({ phase: "end", item });
+    await this.eventProjection.emitNormalizedToolItemEvent({ phase: "result", item });
     this.completedItemIds.add(item.id);
   }
 
   private isCurrentTurnSnapshotItem(item: CodexThreadItem): boolean {
     const itemTurnId = readItemString(item, "turnId");
     return itemTurnId === undefined || itemTurnId === this.turnId;
-  }
-
-  private handleOutputDelta(params: JsonObject, toolName: string): void {
-    const itemId = readString(params, "itemId");
-    const delta = readString(params, "delta");
-    if (!itemId || !delta) {
-      return;
-    }
-    const storedOutput = this.toolOutput.append(itemId, delta);
-    this.rememberToolProgressEcho(itemId, {
-      displayText: storedOutput.text,
-      rawLength: storedOutput.normalizedLength,
-      rawPrefix: storedOutput.rawPrefix,
-      streamedDisplay: true,
-    });
-    if (!this.shouldEmitToolOutput()) {
-      return;
-    }
-    if (
-      this.transcriptToolProgressSuppressedIds.has(itemId) ||
-      !shouldEmitTranscriptToolProgress(toolName, this.toolTranscriptArgumentsById.get(itemId))
-    ) {
-      return;
-    }
-    const state = this.toolResultOutputDeltaState.get(itemId) ?? {
-      chars: 0,
-      messages: 0,
-      truncated: false,
-    };
-    if (state.truncated) {
-      return;
-    }
-    const remainingChars = Math.max(0, TOOL_PROGRESS_OUTPUT_MAX_CHARS - state.chars);
-    const remainingMessages = Math.max(0, MAX_TOOL_OUTPUT_DELTA_MESSAGES_PER_ITEM - state.messages);
-    if (remainingChars === 0 || remainingMessages === 0) {
-      state.truncated = true;
-      this.toolResultOutputDeltaState.set(itemId, state);
-      this.emitToolResultMessage({
-        itemId,
-        text: formatToolOutput(toolName, undefined, "(output truncated)"),
-      });
-      return;
-    }
-    const chunk = delta.length > remainingChars ? truncateUtf16Safe(delta, remainingChars) : delta;
-    state.chars += chunk.length;
-    state.messages += 1;
-    const reachedLimit =
-      delta.length > remainingChars ||
-      state.chars >= TOOL_PROGRESS_OUTPUT_MAX_CHARS ||
-      state.messages >= MAX_TOOL_OUTPUT_DELTA_MESSAGES_PER_ITEM;
-    if (reachedLimit) {
-      state.truncated = true;
-    }
-    this.toolResultOutputDeltaState.set(itemId, state);
-    this.toolResultOutputStreamedItemIds.add(itemId);
-    this.emitToolResultMessage({
-      itemId,
-      text: formatToolOutput(
-        toolName,
-        undefined,
-        reachedLimit ? `${chunk}\n...(truncated)...` : chunk,
-      ),
-    });
   }
 
   private async handleRawResponseItemCompleted(params: JsonObject): Promise<void> {
@@ -914,180 +651,7 @@ export class CodexAppServerEventProjector {
     // Project protocol state before media persistence yields. Notifications may overlap,
     // so delayed image I/O must not consume assistant-echo state from a newer item.
     this.assistantProjection.handleRawResponseItemCompleted(item, this.activeItemIds);
-    await this.recordRawGeneratedImageMedia(item);
-  }
-
-  private recordNativeGeneratedMedia(item: CodexThreadItem | undefined): void {
-    if (item?.type !== "imageGeneration") {
-      return;
-    }
-    const savedPath = readItemString(item, "savedPath")?.trim();
-    if (savedPath) {
-      this.recordNativeGeneratedMediaUrl({
-        itemId: item.id,
-        mediaUrl: savedPath,
-      });
-    }
-  }
-
-  private async recordRawGeneratedImageMedia(item: JsonObject): Promise<void> {
-    if (readString(item, "type") !== "image_generation_call") {
-      return;
-    }
-    const result = readString(item, "result");
-    if (!result) {
-      return;
-    }
-    const itemId = readString(item, "id") ?? `raw-image-${this.nativeGeneratedMediaItemIds.size}`;
-    this.nativeGeneratedMediaItemIds.add(itemId);
-    const maxBytes = resolveGeneratedImageMaxBytes(this.params.config);
-    const estimatedDecodedBytes = estimateBase64DecodedBytes(result);
-    if (estimatedDecodedBytes !== undefined && estimatedDecodedBytes > maxBytes) {
-      embeddedAgentLog.warn("codex app-server raw image generation result exceeds media limit", {
-        itemId,
-        estimatedDecodedBytes,
-        maxBytes,
-      });
-      return;
-    }
-    const asset = generatedImageAssetFromBase64({
-      base64: result,
-      index: this.nativeGeneratedMediaItemIds.size,
-      revisedPrompt: readString(item, "revised_prompt") ?? readString(item, "revisedPrompt"),
-      fileNamePrefix: "codex-image-generation",
-      sniffMimeType: true,
-    });
-    if (!asset) {
-      return;
-    }
-    try {
-      const saved = await saveMediaBuffer(
-        asset.buffer,
-        asset.mimeType,
-        GENERATED_IMAGE_MEDIA_SUBDIR,
-        maxBytes,
-        asset.fileName,
-      );
-      this.recordNativeGeneratedMediaUrl({
-        itemId,
-        mediaUrl: saved.path,
-        // The typed savedPath may belong to a remote app-server host. Always
-        // prefer the copy persisted into this gateway's managed media root.
-        replaceExisting: true,
-      });
-    } catch (error) {
-      embeddedAgentLog.warn("codex app-server raw image generation result save failed", {
-        itemId,
-        error,
-      });
-    }
-  }
-
-  private recordNativeGeneratedMediaUrl(params: {
-    itemId: string;
-    mediaUrl: string;
-    replaceExisting?: boolean;
-  }): void {
-    if (
-      this.nativeGeneratedMediaUrlsByItemId.has(params.itemId) &&
-      params.replaceExisting !== true
-    ) {
-      this.nativeGeneratedMediaItemIds.add(params.itemId);
-      return;
-    }
-    this.nativeGeneratedMediaUrlsByItemId.set(params.itemId, params.mediaUrl);
-    this.nativeGeneratedMediaItemIds.add(params.itemId);
-  }
-
-  private buildToolMediaUrls(toolTelemetry: CodexAppServerToolTelemetry): string[] | undefined {
-    const mediaUrls = new Set(
-      toolTelemetry.toolMediaUrls?.map((url) => url.trim()).filter(Boolean) ?? [],
-    );
-    if ((toolTelemetry.messagingToolSentMediaUrls?.length ?? 0) === 0) {
-      for (const mediaUrl of this.nativeGeneratedMediaUrlsByItemId.values()) {
-        mediaUrls.add(mediaUrl);
-      }
-    }
-    return mediaUrls.size > 0 ? [...mediaUrls] : toolTelemetry.toolMediaUrls;
-  }
-
-  private emitStandardItemEvent(params: {
-    phase: "start" | "end";
-    item: CodexThreadItem | undefined;
-  }): void {
-    const { item } = params;
-    if (!item) {
-      return;
-    }
-    const kind = itemKind(item);
-    if (!kind) {
-      return;
-    }
-    const meta = itemMeta(item, this.toolProgressDetailMode());
-    const suppressChannelProgress = shouldSuppressChannelProgressForItem(item);
-    this.emitAgentEvent({
-      stream: "item",
-      data: {
-        itemId: item.id,
-        phase: params.phase,
-        kind,
-        title: itemTitle(item),
-        status: params.phase === "start" ? "running" : itemStatus(item),
-        ...(itemName(item) ? { name: itemName(item) } : {}),
-        ...(meta ? { meta } : {}),
-        ...(suppressChannelProgress ? { suppressChannelProgress: true } : {}),
-      },
-    });
-  }
-
-  private async emitNormalizedToolItemEvent(params: {
-    phase: "start" | "result";
-    item: CodexThreadItem | undefined;
-  }): Promise<void> {
-    const { item } = params;
-    if (!item || !shouldSynthesizeToolProgressForItem(item)) {
-      return;
-    }
-    const name = itemName(item);
-    if (!name) {
-      return;
-    }
-    const status = params.phase === "result" ? itemStatus(item) : "running";
-    const args = itemToolArgs(item);
-    const meta = itemMeta(item, this.toolProgressDetailMode());
-    this.recordToolTrajectoryEvent({ phase: params.phase, item, name, args, status });
-    if (params.phase === "result") {
-      this.recordNativeToolError({ item, name, meta, status });
-    }
-    if (!shouldEmitTranscriptToolProgress(name, args)) {
-      if (params.phase === "result") {
-        this.emitAfterToolCallObservation(item);
-        await this.options.onNativeToolResultRecorded?.();
-      }
-      return;
-    }
-    this.emitAgentEvent({
-      stream: "tool",
-      data: {
-        phase: params.phase,
-        name,
-        itemId: item.id,
-        toolCallId: item.id,
-        ...(meta ? { meta } : {}),
-        ...(params.phase === "start" && args ? { args } : {}),
-        ...(params.phase === "result"
-          ? {
-              status,
-              isError: isNonSuccessItemStatus(status),
-              ...itemToolResult(item),
-            }
-          : {}),
-      },
-    });
-    if (params.phase === "result") {
-      this.emitAfterToolCallObservation(item);
-      await this.options.onNativeToolResultRecorded?.();
-    }
+    await this.generatedMediaProjection.recordRaw(item);
   }
 
   private clearTerminalPresentationForNativeItem(item: CodexThreadItem | undefined): void {
@@ -1107,459 +671,6 @@ export class CodexAppServerEventProjector {
       ...(toolCallOrdinal !== undefined ? { toolCallOrdinal } : {}),
       terminalPresentation: undefined,
       presentationOnly: true,
-    });
-  }
-
-  private recordNativeToolError(params: {
-    item: CodexThreadItem;
-    name: string;
-    meta?: string;
-    status: ReturnType<typeof itemStatus>;
-  }): void {
-    if (!isNonSuccessItemStatus(params.status)) {
-      if (!this.lastNativeToolError) {
-        return;
-      }
-      if (!this.lastNativeToolError.mutatingAction) {
-        this.lastNativeToolError = undefined;
-        return;
-      }
-      const actionFingerprint = nativeToolActionFingerprint(params.item);
-      if (
-        this.lastNativeToolError.actionFingerprint &&
-        actionFingerprint &&
-        this.lastNativeToolError.actionFingerprint === actionFingerprint
-      ) {
-        this.lastNativeToolError = undefined;
-      }
-      return;
-    }
-    const error = itemToolError(params.item, params.status, this.toolOutput.textByItem);
-    const actionFingerprint = nativeToolActionFingerprint(params.item);
-    this.lastNativeToolError = {
-      toolName: params.name,
-      ...(params.meta ? { meta: params.meta } : {}),
-      ...(error ? { error } : {}),
-      ...(isMutatingNativeToolItem(params.item) ? { mutatingAction: true } : {}),
-      ...(actionFingerprint ? { actionFingerprint } : {}),
-    };
-  }
-
-  private recordToolTrajectoryEvent(params: {
-    phase: "start" | "result";
-    item: CodexThreadItem;
-    name: string;
-    args?: Record<string, unknown>;
-    status: ReturnType<typeof itemStatus>;
-  }): void {
-    if (params.phase === "start") {
-      this.toolTrajectoryCallIds.add(params.item.id);
-      this.toolTrajectoryNamesById.set(params.item.id, params.name);
-      this.toolTrajectoryItemsById.set(params.item.id, params.item);
-      this.options.trajectoryRecorder?.recordEvent("tool.call", {
-        threadId: this.threadId,
-        turnId: this.turnId,
-        itemId: params.item.id,
-        toolCallId: params.item.id,
-        name: params.name,
-        arguments: params.args,
-      });
-      return;
-    }
-    this.toolTrajectoryResultIds.add(params.item.id);
-    const toolResult = itemToolResult(params.item).result;
-    const output = itemOutputText(params.item, this.toolOutput.textByItem);
-    this.options.trajectoryRecorder?.recordEvent("tool.result", {
-      threadId: this.threadId,
-      turnId: this.turnId,
-      itemId: params.item.id,
-      toolCallId: params.item.id,
-      name: params.name,
-      status: params.status,
-      isError: isNonSuccessItemStatus(params.status),
-      ...(toolResult ? { result: toolResult } : {}),
-      ...(output ? { output } : {}),
-    });
-  }
-
-  private emitAfterToolCallObservation(item: CodexThreadItem): void {
-    if (!this.shouldEmitAfterToolCallObservation(item)) {
-      return;
-    }
-    const name = itemName(item);
-    if (!name) {
-      return;
-    }
-    const status = itemStatus(item);
-    if (status === "running") {
-      return;
-    }
-    this.afterToolCallObservedItemIds.add(item.id);
-    const result = itemToolResult(item).result;
-    const error = itemToolError(item, status, this.toolOutput.textByItem);
-    const startedAt = resolveStartedAtFromDurationMs(item.durationMs);
-    const hookParams = {
-      toolName: name,
-      toolCallId: item.id,
-      runId: this.params.runId,
-      agentId: this.params.agentId,
-      sessionId: this.params.sessionId,
-      sessionKey: this.params.sessionKey,
-      startArgs: itemToolArgs(item) ?? {},
-      ...(result !== undefined ? { result } : {}),
-      ...(error ? { error } : {}),
-      ...(startedAt !== undefined ? { startedAt } : {}),
-    };
-    setImmediate(() => {
-      void runAgentHarnessAfterToolCallHook(hookParams);
-    });
-  }
-
-  private shouldEmitAfterToolCallObservation(item: CodexThreadItem): boolean {
-    if (
-      !shouldSynthesizeToolProgressForItem(item) ||
-      this.afterToolCallObservedItemIds.has(item.id)
-    ) {
-      return false;
-    }
-    if (this.options.nativePostToolUseRelayEnabled && isNativePostToolUseRelayItem(item)) {
-      return false;
-    }
-    return true;
-  }
-
-  private emitToolResultSummary(item: CodexThreadItem | undefined): void {
-    if (!item || !this.params.onToolResult || !this.shouldEmitToolResult()) {
-      return;
-    }
-    const itemId = item.id;
-    if (this.toolResultSummaryItemIds.has(itemId)) {
-      return;
-    }
-    const toolName = itemName(item);
-    if (!toolName) {
-      return;
-    }
-    if (!shouldEmitTranscriptToolProgress(toolName, itemToolArgs(item))) {
-      return;
-    }
-    this.toolResultSummaryItemIds.add(itemId);
-    const meta = itemMeta(item, this.toolProgressDetailMode());
-    this.emitToolResultMessage({
-      itemId,
-      text: formatToolSummary(toolName, meta),
-    });
-  }
-
-  private emitToolResultOutput(item: CodexThreadItem | undefined): void {
-    if (!item || !this.params.onToolResult || !this.shouldEmitToolOutput()) {
-      return;
-    }
-    const itemId = item.id;
-    if (this.toolResultOutputItemIds.has(itemId)) {
-      return;
-    }
-    if (this.toolResultOutputStreamedItemIds.has(itemId)) {
-      return;
-    }
-    const toolName = itemName(item);
-    const output = itemOutputText(item, this.toolOutput.textByItem);
-    if (!toolName || !output) {
-      return;
-    }
-    if (!shouldEmitTranscriptToolProgress(toolName, itemToolArgs(item))) {
-      return;
-    }
-    this.emitToolResultMessage({
-      itemId,
-      text: formatToolOutput(toolName, itemMeta(item, this.toolProgressDetailMode()), output),
-      finalOutput: true,
-      isError: isNonSuccessItemStatus(itemStatus(item)),
-    });
-  }
-
-  private emitToolResultMessage(params: {
-    itemId: string;
-    text: string;
-    finalOutput?: boolean;
-    isError?: boolean;
-  }): void {
-    const rawText = params.text.trim();
-    const text = truncateToolTranscriptText(rawText);
-    if (!text) {
-      return;
-    }
-    this.rememberToolProgressEcho(params.itemId, { displayText: text, rawText });
-    if (params.finalOutput) {
-      this.toolResultOutputItemIds.add(params.itemId);
-    }
-    try {
-      void Promise.resolve(
-        this.params.onToolResult?.({
-          text,
-          ...(params.isError === true ? { isError: true } : {}),
-        }),
-      ).catch(() => {
-        // Tool progress delivery is best-effort and should not affect the turn.
-      });
-    } catch {
-      // Tool progress delivery is best-effort and should not affect the turn.
-    }
-  }
-
-  private shouldEmitToolResult(): boolean {
-    return typeof this.params.shouldEmitToolResult === "function"
-      ? this.params.shouldEmitToolResult()
-      : this.params.verboseLevel === "on" || this.params.verboseLevel === "full";
-  }
-
-  private shouldEmitToolOutput(): boolean {
-    return typeof this.params.shouldEmitToolOutput === "function"
-      ? this.params.shouldEmitToolOutput()
-      : this.params.verboseLevel === "full";
-  }
-
-  private toolProgressDetailMode(): ToolProgressDetailMode {
-    return resolveCodexToolProgressDetailMode(this.params.toolProgressDetail);
-  }
-
-  private recordToolMeta(item: CodexThreadItem | undefined): void {
-    if (!item) {
-      return;
-    }
-    if (isSideEffectingNativeToolItem(item)) {
-      this.sideEffectingToolItemIds.add(item.id);
-    } else {
-      this.sideEffectingToolItemIds.delete(item.id);
-    }
-    const toolName = itemName(item);
-    if (!toolName) {
-      return;
-    }
-    const meta = itemMeta(item, this.toolProgressDetailMode());
-    const status = itemStatus(item);
-    const existing = this.toolMetas.get(item.id);
-    this.toolMetas.set(item.id, {
-      toolName,
-      ...(meta ? { meta } : {}),
-      ...(existing?.asyncStarted ? { asyncStarted: true } : {}),
-      ...(status !== "running" && isNonSuccessItemStatus(status) ? { isError: true } : {}),
-    });
-  }
-
-  private recordNativeToolTranscriptCall(item: CodexThreadItem | undefined): void {
-    if (!item || !shouldRecordNativeToolTranscript(item)) {
-      return;
-    }
-    const name = itemName(item);
-    if (!name) {
-      return;
-    }
-    this.recordToolTranscriptCall({
-      id: item.id,
-      name,
-      arguments: itemToolArgs(item),
-    });
-  }
-
-  private recordNativeToolTranscriptResult(item: CodexThreadItem | undefined): void {
-    if (!item || !shouldRecordNativeToolTranscript(item)) {
-      return;
-    }
-    const name = itemName(item);
-    if (!name) {
-      return;
-    }
-    this.recordToolTranscriptResult({
-      id: item.id,
-      name,
-      text: itemTranscriptResultText(item, this.toolOutput.textByItem),
-      isError: isNonSuccessItemStatus(itemStatus(item)),
-    });
-  }
-
-  private recordToolTranscriptCall(params: ToolTranscriptCallInput): void {
-    if (!params.id || !params.name || this.toolTranscriptCallIds.has(params.id)) {
-      return;
-    }
-    this.toolTranscriptCallIds.add(params.id);
-    this.toolTranscriptNamesById.set(params.id, params.name);
-    this.toolTranscriptArgumentsById.set(params.id, params.arguments);
-    if (!shouldEmitTranscriptToolProgress(params.name, params.arguments)) {
-      this.transcriptToolProgressSuppressedIds.add(params.id);
-    } else {
-      this.transcriptToolProgressSuppressedIds.delete(params.id);
-    }
-    this.emitTranscriptToolCallProgress(params);
-    this.toolTranscriptMessages.push(
-      attachCodexMirrorIdentity(
-        this.createToolCallMessage(params),
-        `${this.turnId}:tool:${params.id}:call`,
-      ),
-    );
-  }
-
-  private recordToolTranscriptResult(params: ToolTranscriptResultInput): void {
-    if (!params.id || !params.name || this.toolTranscriptResultIds.has(params.id)) {
-      return;
-    }
-    this.toolTranscriptResultIds.add(params.id);
-    this.emitTranscriptToolResultProgress(params);
-    this.toolTranscriptMessages.push(
-      attachCodexMirrorIdentity(
-        this.createToolResultMessage(params),
-        `${this.turnId}:tool:${params.id}:result`,
-      ),
-    );
-  }
-
-  private synthesizeMissingToolResults(params: {
-    synthesize: boolean;
-    recordPromptError: boolean;
-  }): void {
-    if (!params.synthesize) {
-      return;
-    }
-    const missingTranscriptIds = [...this.toolTranscriptCallIds].filter(
-      (id) => !this.toolTranscriptResultIds.has(id),
-    );
-    const missingTrajectoryIds = [...this.toolTrajectoryCallIds].filter(
-      (id) => !this.toolTrajectoryResultIds.has(id),
-    );
-    if (missingTranscriptIds.length === 0 && missingTrajectoryIds.length === 0) {
-      return;
-    }
-
-    for (const id of missingTranscriptIds) {
-      const name = this.toolTranscriptNamesById.get(id) ?? this.toolTrajectoryNamesById.get(id);
-      if (!name) {
-        continue;
-      }
-      this.recordToolTranscriptResult({
-        id,
-        name,
-        text: formatMissingToolResultError({ id, name }),
-        isError: true,
-      });
-    }
-
-    for (const id of missingTrajectoryIds) {
-      const name = this.toolTrajectoryNamesById.get(id) ?? this.toolTranscriptNamesById.get(id);
-      if (!name) {
-        continue;
-      }
-      this.toolTrajectoryResultIds.add(id);
-      const text = formatMissingToolResultError({ id, name });
-      this.options.trajectoryRecorder?.recordEvent("tool.result", {
-        threadId: this.threadId,
-        turnId: this.turnId,
-        itemId: id,
-        toolCallId: id,
-        name,
-        status: "failed",
-        isError: true,
-        result: { status: "failed", reason: "missing_tool_result" },
-        output: text,
-      });
-    }
-
-    if (!params.recordPromptError) {
-      const firstMissingId =
-        missingTranscriptIds.find((id) => {
-          const name = this.toolTranscriptNamesById.get(id) ?? this.toolTrajectoryNamesById.get(id);
-          return Boolean(name);
-        }) ??
-        missingTrajectoryIds.find((id) => {
-          const name = this.toolTrajectoryNamesById.get(id) ?? this.toolTranscriptNamesById.get(id);
-          return Boolean(name);
-        });
-      if (firstMissingId) {
-        const name =
-          this.toolTranscriptNamesById.get(firstMissingId) ??
-          this.toolTrajectoryNamesById.get(firstMissingId);
-        if (name) {
-          const item = this.toolTrajectoryItemsById.get(firstMissingId);
-          const meta = item
-            ? itemMeta(item, this.toolProgressDetailMode())
-            : this.toolMetas.get(firstMissingId)?.meta;
-          const actionFingerprint = item ? nativeToolActionFingerprint(item) : undefined;
-          this.lastNativeToolError = {
-            toolName: name,
-            ...(meta ? { meta } : {}),
-            error: formatMissingToolResultError({ id: firstMissingId, name }),
-            ...(item && isMutatingNativeToolItem(item) ? { mutatingAction: true } : {}),
-            ...(actionFingerprint ? { actionFingerprint } : {}),
-          };
-        }
-      }
-      return;
-    }
-    const missingCount = new Set([...missingTranscriptIds, ...missingTrajectoryIds]).size;
-    this.synthesizedMissingToolResultError =
-      missingCount === 1
-        ? MISSING_TOOL_RESULT_ERROR
-        : `${MISSING_TOOL_RESULT_ERROR} missingToolResultCount=${missingCount}`;
-    this.promptErrorSource = this.promptErrorSource ?? "prompt";
-  }
-
-  private emitTranscriptToolCallProgress(params: ToolTranscriptCallInput): void {
-    if (!shouldEmitTranscriptToolProgress(params.name, params.arguments)) {
-      return;
-    }
-    this.transcriptToolProgressCallIds.add(params.id);
-    const args = normalizeToolTranscriptArguments(params.arguments);
-    const meta = inferToolMetaFromArgs(params.name, args, {
-      detailMode: this.toolProgressDetailMode(),
-    });
-    if (
-      !this.params.onToolResult ||
-      !this.shouldEmitToolResult() ||
-      this.toolResultSummaryItemIds.has(params.id) ||
-      this.toolResultOutputStreamedItemIds.has(params.id)
-    ) {
-      return;
-    }
-    this.toolResultSummaryItemIds.add(params.id);
-    this.emitToolResultMessage({
-      itemId: params.id,
-      text: formatToolSummary(params.name, meta),
-    });
-  }
-
-  private emitTranscriptToolResultProgress(params: ToolTranscriptResultInput): void {
-    if (
-      this.transcriptToolProgressSuppressedIds.has(params.id) ||
-      !shouldEmitTranscriptToolProgress(
-        params.name,
-        this.toolTranscriptArgumentsById.get(params.id),
-      )
-    ) {
-      return;
-    }
-    if (!this.transcriptToolProgressCallIds.has(params.id)) {
-      this.emitTranscriptToolCallProgress({
-        id: params.id,
-        name: params.name,
-        arguments: {},
-      });
-    }
-    if (
-      !this.params.onToolResult ||
-      !this.shouldEmitToolOutput() ||
-      this.toolResultOutputItemIds.has(params.id) ||
-      this.toolResultOutputStreamedItemIds.has(params.id)
-    ) {
-      return;
-    }
-    const text = params.text?.trim();
-    if (!text) {
-      return;
-    }
-    this.emitToolResultMessage({
-      itemId: params.id,
-      text: formatToolOutput(params.name, undefined, text),
-      finalOutput: true,
-      isError: params.isError,
     });
   }
 
@@ -1598,159 +709,6 @@ export class CodexAppServerEventProjector {
     }
   }
 
-  private matchesToolProgressEcho(text: string): boolean {
-    for (const state of this.toolProgressEchoesByItem.values()) {
-      if (state.streamedDisplayText === text) {
-        return true;
-      }
-      if (state.displayTexts.includes(text)) {
-        return true;
-      }
-      if (
-        state.streamedRawSignature &&
-        text.length === state.streamedRawSignature.length &&
-        text.startsWith(state.streamedRawSignature.prefix)
-      ) {
-        return true;
-      }
-      for (const signature of state.rawSignatures) {
-        if (text.length === signature.length && text.startsWith(signature.prefix)) {
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  private rememberToolProgressEcho(
-    itemId: string,
-    signature: {
-      displayText?: string;
-      rawText?: string;
-      rawLength?: number;
-      rawPrefix?: string;
-      streamedDisplay?: boolean;
-    },
-  ): void {
-    if (!itemId) {
-      return;
-    }
-    const existing = this.toolProgressEchoesByItem.get(itemId) ?? {
-      displayTexts: [],
-      rawSignatures: [],
-    };
-    const displayText = signature.displayText?.trim();
-    if (displayText) {
-      if (signature.streamedDisplay) {
-        existing.streamedDisplayText = displayText;
-      } else if (!existing.displayTexts.includes(displayText)) {
-        if (existing.displayTexts.length >= TOOL_PROGRESS_ECHO_SIGNATURE_CAP) {
-          existing.displayTexts.shift();
-        }
-        existing.displayTexts.push(displayText);
-      }
-    }
-    const rawText = signature.rawText?.trim();
-    const rawLength = signature.rawLength ?? rawText?.length;
-    const rawPrefix = signature.rawPrefix?.trim() ?? rawText;
-    if (
-      rawLength !== undefined &&
-      rawPrefix &&
-      rawPrefix.length >= TOOL_PROGRESS_ECHO_PREFIX_MIN_CHARS
-    ) {
-      const next: ToolProgressRawSignature = {
-        length: rawLength,
-        prefix: rawPrefix.slice(0, TOOL_TRANSCRIPT_OUTPUT_MAX_CHARS),
-      };
-      if (signature.streamedDisplay) {
-        // Stream accumulation is one logical shape; replace the dedicated slot only.
-        existing.streamedRawSignature = next;
-      } else {
-        const matchIndex = existing.rawSignatures.findIndex(
-          (entry) => entry.prefix === next.prefix,
-        );
-        if (matchIndex >= 0) {
-          existing.rawSignatures[matchIndex] = next;
-        } else {
-          if (existing.rawSignatures.length >= TOOL_PROGRESS_ECHO_SIGNATURE_CAP) {
-            existing.rawSignatures.shift();
-          }
-          existing.rawSignatures.push(next);
-        }
-      }
-    }
-    this.toolProgressEchoesByItem.set(itemId, existing);
-  }
-
-  private rememberCommandAggregateOutputEcho(item: CodexThreadItem | undefined): void {
-    if (item?.type !== "commandExecution" || typeof item.aggregatedOutput !== "string") {
-      return;
-    }
-    const signature = toolOutputRawEchoSignature(item.aggregatedOutput);
-    if (!signature) {
-      return;
-    }
-    this.rememberToolProgressEcho(item.id, signature);
-  }
-
-  private async readMirroredSessionMessages(): Promise<AgentMessage[]> {
-    return (
-      (await readCodexMirroredSessionHistoryMessages({
-        agentId: this.params.agentId,
-        sessionFile: this.params.sessionFile,
-        sessionId: this.params.sessionId,
-        sessionKey: this.params.sessionKey,
-      })) ?? []
-    );
-  }
-
-  private createToolCallMessage(params: ToolTranscriptCallInput): AgentMessage {
-    const args = normalizeToolTranscriptArguments(params.arguments);
-    const attribution = resolveCodexLocalRuntimeAttribution(this.params);
-    return {
-      role: "assistant",
-      content: [
-        {
-          type: "toolCall",
-          id: params.id,
-          name: params.name,
-          arguments: args,
-          input: args,
-        },
-      ],
-      api: attribution.api ?? "openai-chatgpt-responses",
-      provider: attribution.provider,
-      model: this.params.modelId,
-      usage: ZERO_USAGE,
-      stopReason: "toolUse",
-      timestamp: Date.now(),
-    } as unknown as AgentMessage;
-  }
-
-  private createToolResultMessage(params: ToolTranscriptResultInput): AgentMessage {
-    const text = truncateToolTranscriptText(params.text?.trim() || toolResultStatusText(params));
-    return {
-      role: "toolResult",
-      toolCallId: params.id,
-      toolName: params.name,
-      isError: params.isError,
-      content: [
-        {
-          type: "toolResult",
-          id: params.id,
-          name: params.name,
-          toolName: params.name,
-          toolCallId: params.id,
-          toolUseId: params.id,
-          tool_use_id: params.id,
-          content: text,
-          text,
-        },
-      ],
-      timestamp: Date.now(),
-    } as unknown as AgentMessage;
-  }
-
   private isNotificationForTurn(params: JsonObject): boolean {
     const threadId = readCodexNotificationThreadId(params);
     const turnId = readCodexNotificationTurnId(params);
@@ -1766,61 +724,4 @@ export class CodexAppServerEventProjector {
 
 function isHookNotificationMethod(method: string): method is "hook/started" | "hook/completed" {
   return method === "hook/started" || method === "hook/completed";
-}
-
-function estimateBase64DecodedBytes(base64: string): number | undefined {
-  let nonWhitespaceLength = 0;
-  let previousCode = -1;
-  let lastCode = -1;
-  for (let i = 0; i < base64.length; i += 1) {
-    const code = base64.charCodeAt(i);
-    if (isBase64WhitespaceCode(code)) {
-      continue;
-    }
-    nonWhitespaceLength += 1;
-    previousCode = lastCode;
-    lastCode = code;
-  }
-  if (nonWhitespaceLength === 0) {
-    return undefined;
-  }
-  const equalsCode = "=".charCodeAt(0);
-  const padding = lastCode === equalsCode ? (previousCode === equalsCode ? 2 : 1) : 0;
-  return Math.max(0, Math.floor((nonWhitespaceLength * 3) / 4) - padding);
-}
-
-function isBase64WhitespaceCode(code: number): boolean {
-  return code === 0x20 || code === 0x09 || code === 0x0a || code === 0x0d;
-}
-
-function resolveGeneratedImageMaxBytes(config: EmbeddedRunAttemptParams["config"]): number {
-  const configured = config?.agents?.defaults?.mediaMaxMb;
-  if (typeof configured === "number" && Number.isFinite(configured) && configured > 0) {
-    return Math.floor(configured * BYTES_PER_MB);
-  }
-  return DEFAULT_GENERATED_IMAGE_MAX_BYTES;
-}
-
-function resolveStartedAtFromDurationMs(durationMs: unknown): number | undefined {
-  if (typeof durationMs !== "number" || !Number.isFinite(durationMs)) {
-    return undefined;
-  }
-  return asDateTimestampMs(Date.now() - Math.max(0, durationMs));
-}
-
-function normalizeCodexTokenUsage(record: JsonObject): ReturnType<typeof normalizeUsage> {
-  // v2 TokenUsageBreakdown. inputTokens includes cached input; OpenClaw usage
-  // tracks uncached input and cache reads separately.
-  const inputTokens = readNumber(record, "inputTokens");
-  const cacheRead = readNumber(record, "cachedInputTokens");
-  const input =
-    inputTokens !== undefined && cacheRead !== undefined
-      ? Math.max(0, inputTokens - cacheRead)
-      : inputTokens;
-  return normalizeUsage({
-    input,
-    output: readNumber(record, "outputTokens"),
-    cacheRead,
-    total: readNumber(record, "totalTokens"),
-  });
 }
